// src/app/interfaces/questions.ts
export class Question {
    username: string;
    Qname: string;
    Qdescrib: string;
    question1: string;
    question2: string;
    question3: string;
    question4: string;
    question5: string;
    question6: string;
    question7: string;
    question8: string;
    question9: string;
    question10: string;
    question11: string;
    question12: string;
    constructor() {
        this.question1 = '';
        this.username = '';
        this.Qname = '';
        this.Qdescrib = '';
        this.question1 = '';
        this.question2 = '';
        this.question3 = '';
        this.question4 = '';
        this.question5 = '';
        this.question6 = '';
        this.question7 = '';
        this.question8 = '';
        this.question9 = '';
        this.question10 = '';
        this.question11 = '';
        this.question12 = '';
    }

}